requirements.txt
python-telegram-bot==13.2
requests
instaloader
python-dotenv
pokepy